Octrees are now generated directly by Geometry processing instead of using an intermediate plist representation. Unnecessary copies of octree data eliminated.
git-svn-id: http://svn.berlios.de/svnroot/repos/oolite-linux/trunk@5363 127b21dd-08f5-0310-b4b7-95ae10353056

+++ src/Core/Geometry.h
@@ -46,6 +46,6 @@ MA 02110-1301, USA.
 
 - (void) addTriangle:(Triangle) tri;
 
-- (Octree *) findOctreeToDepth:(int)depth;
+- (Octree *) findOctreeToDepth:(NSUInteger)depth;
 
 @end
